+++ messaging-core/src/test/java/com/nabla/sdk/messaging/core/data/message/GqlConversationContentDataSourceTest.kt
@@ -254,6 +254,7 @@ class GqlConversationContentDataSourceTest : BaseCoroutineTest() {
             coroutineScope = scope,
             apolloClient = apolloClient,
             mapper = GqlMapper(logger, localConversationDataSource, coreGqlMapper),
+            localConversationDataSource = localConversationDataSource,
         )
     }
 }