+++ CHANGELOG.md
@@ -16,6 +16,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Fixed
 
+- Messaging: Fix error when deleting a message on a freshly created draft conversation (i.e. with Local id).
 - Fix a crash on Android 6 and 7 when taking a picture with the camera in messaging UI module.
 
 ## [1.0-alpha17] - 2022-10-19